requirements.txt
tensorflow==2.20.0
numpy==2.4.2
matplotlib==3.10.8
scikit-learn==1.8.0
seaborn==0.13.2
Pillow==12.1.1
flask==3.1.3
